+++ src/packages/piral-core/src/modules/events.ts
@@ -1,29 +1,35 @@
-import { Listener, EventEmitter } from '../types';
-import { appendItem, excludeItem } from '../utils';
+import { EventEmitter } from '../types';
 
-export interface EventListeners {
-  [event: string]: Array<Listener<any>>;
+export type EventListeners = Array<[any, any]>;
+
+function nameOf(type: string | number) {
+  return `piral-${type}`;
 }
 
 export function createListener(): EventEmitter {
-  const eventListeners: EventListeners = {};
+  const eventListeners: EventListeners = [];
 
   return {
     on(type, callback) {
-      eventListeners[type] = appendItem(eventListeners[type], callback);
+      const listener = (ev: CustomEvent) => callback(ev.detail);
+      document.body.addEventListener(nameOf(type), listener);
+      eventListeners.push([callback, listener]);
       return this;
     },
     off(type, callback) {
-      eventListeners[type] = excludeItem(eventListeners[type], callback);
+      const [listener] = eventListeners.filter(m => m[0] === callback);
+
+      if (listener) {
+        document.body.removeEventListener(nameOf(type), listener[1]);
+        eventListeners.splice(eventListeners.indexOf(listener), 1);
+      }
+
       return this;
     },
     emit(type, arg) {
-      const callbacks = eventListeners[type] || [];
-
-      for (const callback of callbacks) {
-        setTimeout(() => callback.call(this, arg), 0);
-      }
-
+      const ce = document.createEvent('CustomEvent');
+      ce.initCustomEvent(nameOf(type), false, false, arg);
+      document.body.dispatchEvent(ce);
       return this;
     },
   };