+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Piral Changelog
 
+## 0.5.1 (to be determined)
+
+* Added new optional library `piral-ng`
+* Extended APIs for `registerExtension` and `registerModal`
+* Improved documentation
+* Added more flexibility to `piral build` and `piral debug` commands
+* Improved bundling with splitting, module resolution, and URLs
+
 ## 0.5.0 (June 13, 2019)
 
 * Added more props to the `Layout`